use single native callback in application menu on macOS
This change is intended to address the following issue, sometimes item trigger may decide to update the application menu. So the callback currently in the call stack will be freed with the current callback arena. It leads to pretty strange segfaults in GC code. To address that, we switch to a single native callback for the whole app menu with almost iternal lifetime.